+++ backend/src/main/resources/application.properties
@@ -1,7 +1,7 @@
 spring.datasource.url=jdbc:mariadb://loain.cab269b7tt71.ap-northeast-2.rds.amazonaws.com:3306/loain?serverTimezone=Asia/Seoul&characterEncoding=UTF-8
 spring.datasource.driver-class-name=org.mariadb.jdbc.Driver
 spring.datasource.username=root
-spring.datasource.password=789qwe!!
+spring.datasource.password=smilegate6974
 
 #MyBatis
 mybatis.type-aliases-package=loa.backend.Model
